Refactor code to have endpoints in controllers
Created controllers for endpoints.
Added a controllers to redirect to the root of the site if the
session is over od don't exist.
Renamed things.
Added (simple) configuration steps to the README.
