+++ Procfile
@@ -1,1 +1,1 @@
-web: bundle exec puma -Ilib -Iapp -C config/puma.rb
+web: bundle exec puma -C config/puma.rb
