+++ app/tentacles/controllers.rb
@@ -1,3 +1,7 @@
+require 'tentacles/controllers/authentication'
+require 'tentacles/controllers/repositories'
+require 'tentacles/controllers/pulls'
+
 module Tentacles
   ##
   # Controllers Namespace
