+++ app/tentacles/helpers/authentication.rb
@@ -1,0 +1,11 @@
+module Tentacles
+  module Helpers
+    ##
+    # Authentication helper
+    module Authentication
+      def authenticated?
+        session && session[:access_token]
+      end
+    end
+  end
+end
