+++ app/tentacles/views/repositories.erb
@@ -9,7 +9,7 @@
 
     <h2>Repositories</h2>
     <p>Please select a list of repositories you would like to display the Pull Requests:</p>
-    <form class="user-repos" action="/pull" method="post">
+    <form class="user-repos" action="/pulls" method="post">
     <ul class ="">
       <% repos.each do |repo| %>
       <li class ="">
